Refactor code and add accessor for rps subgraphs
- The utility random function has been changed to indicate it works
on indexes, which makes it clearer that the passed limit is exclusive.
- Add a private property to rps for the number of available choices that
simplifies code.
- `isValidChoice` now queries the graph directly to take advantage of
the lookup speed.
- Handling of the string passed to the rps class should not be the
responsibility of the class itself but rather of a controller module.
- The logic of getting the result is moved to playRound as this is
essentially the one thing this function does.
- The main graph is extended to include both winning and losing outcomes
because this information needs to be fast to lookup and return for drawing
purposes in the future.